Create Solution_Colab_A11_data_comprehension_Kamisha_Godwin.ipynb
I worked on list and dictionary comprehension, converting list to dictionary and vice versa. Then , I also worked on sets, specifically working on symmetric difference and difference between two sets. After that , I also made a function to find the percentage of G and C of every dna in a list of dnas and printing out their respective %GC.